fix standlone move revert
