+++ sample-operators/webpage/src/main/java/io/javaoperatorsdk/operator/sample/IngressDependentResource.java
@@ -1,11 +1,9 @@
-package io.javaoperatorsdk.operator.sample.dependents;
+package io.javaoperatorsdk.operator.sample;
 
 import io.fabric8.kubernetes.api.model.networking.v1.Ingress;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.processing.dependent.kubernetes.CRUKubernetesDependentResource;
 import io.javaoperatorsdk.operator.processing.dependent.kubernetes.KubernetesDependent;
-import io.javaoperatorsdk.operator.sample.WebPage;
-import io.javaoperatorsdk.operator.sample.WebPageManagedDependentsReconciler;
 
 import static io.javaoperatorsdk.operator.sample.Utils.*;
