+++ sample-operators/webpage/src/main/java/io/javaoperatorsdk/operator/sample/Utils.java
@@ -9,7 +9,7 @@ public class Utils {
 
   private Utils() {}
 
-  public static WebPageStatus createStatus(String configMapName) {
+  static WebPageStatus createStatus(String configMapName) {
     WebPageStatus status = new WebPageStatus();
     status.setHtmlConfigMap(configMapName);
     status.setAreWeGood(true);
@@ -17,37 +17,37 @@ public class Utils {
     return status;
   }
 
-  public static String configMapName(WebPage nginx) {
+  static String configMapName(WebPage nginx) {
     return nginx.getMetadata().getName() + "-html";
   }
 
-  public static String deploymentName(WebPage nginx) {
+  static String deploymentName(WebPage nginx) {
     return nginx.getMetadata().getName();
   }
 
-  public static String serviceName(WebPage webPage) {
+  static String serviceName(WebPage webPage) {
     return webPage.getMetadata().getName();
   }
 
-  public static ErrorStatusUpdateControl<WebPage> handleError(WebPage resource, Exception e) {
+  static ErrorStatusUpdateControl<WebPage> handleError(WebPage resource, Exception e) {
     resource.getStatus().setErrorMessage("Error: " + e.getMessage());
     return ErrorStatusUpdateControl.updateStatus(resource);
   }
 
-  public static void simulateErrorIfRequested(WebPage webPage) throws ErrorSimulationException {
+  static void simulateErrorIfRequested(WebPage webPage) throws ErrorSimulationException {
     if (webPage.getSpec().getHtml().contains("error")) {
       // special case just to showcase error if doing a demo
       throw new ErrorSimulationException("Simulating error");
     }
   }
 
-  public static boolean isValidHtml(WebPage webPage) {
+  static boolean isValidHtml(WebPage webPage) {
     // very dummy html validation
     var lowerCaseHtml = webPage.getSpec().getHtml().toLowerCase();
     return lowerCaseHtml.contains("<html>") && lowerCaseHtml.contains("</html>");
   }
 
-  public static WebPage setInvalidHtmlErrorMessage(WebPage webPage) {
+  static WebPage setInvalidHtmlErrorMessage(WebPage webPage) {
     if (webPage.getStatus() == null) {
       webPage.setStatus(new WebPageStatus());
     }
@@ -55,7 +55,7 @@ public class Utils {
     return webPage;
   }
 
-  public static Ingress makeDesiredIngress(WebPage webPage) {
+  static Ingress makeDesiredIngress(WebPage webPage) {
     Ingress ingress = loadYaml(Ingress.class, Utils.class, "ingress.yaml");
     ingress.getMetadata().setName(webPage.getMetadata().getName());
     ingress.getMetadata().setNamespace(webPage.getMetadata().getNamespace());
